Solution to Movie Ticket Booking System

+++ Movie Ticket Booking System@@67877d61dc107486ad5c2751/code.c
@@ -4,11 +4,11 @@
 int main() {
     int x;
     scanf("%d", &x);
-    int rev_standard = 0, rev_prem = 0, rev_vip = 0;
+    float rev_standard = 0, rev_prem = 0, rev_vip = 0;
     
     for (int i = 0; i < x; i++) {
         char movie_name[10], sub[10];
-        int price;
+        float price;
         scanf("%s %s %f", movie_name, sub, &price);
         
         if (strcmp(sub, "Standard") == 0) {
